pyproject.toml
[project]
name = "uisurf-agent"
version = "0.1.0"
description = "A Universal agent capable of reasoning over differnt uis and performing actions on them."
readme = "README.md"
authors = [
    { name = "haruiz", email = "henryruiz22@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk[http-server]>=0.3.25",
    "google-cloud-aiplatform>=1.141.0",
    "mss>=10.1.0",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "termcolor>=3.1.0",
    "typer>=0.16.0",
    "uvicorn>=0.41.0",
]
[project.scripts]
uisurf-agent = "uisurf_agent.cli:app"
uisurf_agent = "uisurf_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
